Add passing the receipe and some little changes

+++ src/main/java/com/netckracker/graph/manager/service/TagServiceImpl.java
@@ -55,11 +55,11 @@ public class TagServiceImpl implements TagService{
     }
 
     @Override
-    public List<ReceipeDto> findByTag(String tagName, int page, int size) {
+    public List<ReceipeDto> findByTag(String tagName, Integer page, Integer size) {
         Tags tag=tagRepository.findByName(tagName);
         if (tag!=null)
         {
-            List<Receipe> receipes=receipeRepository.findByTag(tag.getTagId(), new PageRequest(page, size)).getContent();
+            List<Receipe> receipes=receipeRepository.findByTag(tag.getTagId(), new PageRequest(page.intValue(), size.intValue())).getContent();
             return receipes.stream()
                .map(receipe->convertor.convertReceipeToDto(receipe))
                .collect(Collectors.toList());
@@ -68,8 +68,8 @@ public class TagServiceImpl implements TagService{
     }
 
     @Override
-    public List<TagsDto> findByFirstLetters(String letters, int page, int size) {
-        List<Tags> tags=tagRepository.findFirst10ByNameStartingWith(letters, new PageRequest(page, size)).getContent();
+    public List<TagsDto> findByFirstLetters(String letters, Integer page, Integer size) {
+        List<Tags> tags=tagRepository.findFirst10ByNameStartingWith(letters, new PageRequest(page.intValue(), size.intValue())).getContent();
         return tags.stream()
                .map(tag->convertor.convertTagsToDto(tag))
                .collect(Collectors.toList());
